Split the sql-connection crate into modules
